requirements.txt
aniso8601==4.0.1
atomicwrites==1.2.1
attrs==18.2.0
autopep8==1.4.3
Click==7.0
coverage==4.5.2
Flask==1.0.2
Flask-RESTful==0.3.6
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
more-itertools==4.3.0
nose==1.3.7
pep8==1.7.1
pluggy==0.8.0
py==1.7.0
pycodestyle==2.4.0
pytest==4.0.1
pytest-cov==2.6.0
python-dotenv==0.9.1
pytz==2018.7
six==1.11.0
Werkzeug==0.14.1
